+++ packages/core/src/index.ts
@@ -19,6 +19,10 @@ export type * from "./application/ports/index.js";
 // Use-cases
 export { CreateGoal } from "./application/use-cases/CreateGoal.js";
 export type { CreateGoalInput } from "./application/use-cases/CreateGoal.js";
+export { EditGoal } from "./application/use-cases/EditGoal.js";
+export type { EditGoalInput } from "./application/use-cases/EditGoal.js";
+export { ArchiveGoal } from "./application/use-cases/ArchiveGoal.js";
+export type { ArchiveGoalInput } from "./application/use-cases/ArchiveGoal.js";
 export { CompleteTask } from "./application/use-cases/CompleteTask.js";
 export type { CompleteTaskInput } from "./application/use-cases/CompleteTask.js";
 export { LogExecution } from "./application/use-cases/LogExecution.js";